Move cache cleaner to master node
